cloud-functions/entgeltatlas/requirements.txt
functions-framework==3.*
google-cloud-logging
requests